Types & API Definition fixes:
- make it partly raml1.0-rc2 compliant
- remove map
- linter fix
