--- codegen/fixtures/struct/mapOfCat.txt
@@ -1,10 +1,0 @@
-package main
-
-import ()
-
-type mapOfCat map[string]Cat
-
-func (s mapOfCat) Validate() error {
-
-	return nil
-}
