--- codegen/fixtures/struct/mapOfNumber.txt
@@ -1,10 +1,0 @@
-package main
-
-import ()
-
-type mapOfNumber map[string]float64
-
-func (s mapOfNumber) Validate() error {
-
-	return nil
-}
