--- codegen/fixtures/struct/mapOfString.txt
@@ -1,10 +1,0 @@
-package main
-
-import ()
-
-type mapOfString map[string]string
-
-func (s mapOfString) Validate() error {
-
-	return nil
-}
