--- codegen/fixtures/struct/mapofcats.txt
@@ -1,10 +1,0 @@
-package main
-
-import ()
-
-type mapOfCats map[string]Cat
-
-func (s mapOfCats) Validate() error {
-
-	return nil
-}
