+++ codegen/struct_test.go
@@ -75,15 +75,6 @@ func TestGenerateStructFromRaml(t *testing.T) {
 
 			So(s, ShouldEqual, tmpl)
 
-			// map type
-			s, err = testLoadFile(filepath.Join(targetdir, "mapOfCats.go"))
-			So(err, ShouldBeNil)
-
-			tmpl, err = testLoadFile("./fixtures/struct/mapofcats.txt")
-			So(err, ShouldBeNil)
-
-			So(s, ShouldEqual, tmpl)
-
 			// using map type & testing case sensitive type name
 			s, err = testLoadFile(filepath.Join(targetdir, "petshop.go"))
 			So(err, ShouldBeNil)
@@ -116,33 +107,6 @@ func TestGenerateStructFromRaml(t *testing.T) {
 			So(err, ShouldBeNil)
 
 			tmpl, err = testLoadFile("./fixtures/struct/Specialization.txt")
-			So(err, ShouldBeNil)
-
-			So(s, ShouldEqual, tmpl)
-
-			// Map of cat
-			s, err = testLoadFile(filepath.Join(targetdir, "mapOfCat.go"))
-			So(err, ShouldBeNil)
-
-			tmpl, err = testLoadFile("./fixtures/struct/mapOfCat.txt")
-			So(err, ShouldBeNil)
-
-			So(s, ShouldEqual, tmpl)
-
-			// Map of number
-			s, err = testLoadFile(filepath.Join(targetdir, "mapOfNumber.go"))
-			So(err, ShouldBeNil)
-
-			tmpl, err = testLoadFile("./fixtures/struct/mapOfNumber.txt")
-			So(err, ShouldBeNil)
-
-			So(s, ShouldEqual, tmpl)
-
-			// Map of string
-			s, err = testLoadFile(filepath.Join(targetdir, "mapOfString.go"))
-			So(err, ShouldBeNil)
-
-			tmpl, err = testLoadFile("./fixtures/struct/mapOfString.txt")
 			So(err, ShouldBeNil)
 
 			So(s, ShouldEqual, tmpl)
