+++ codegen/types_test.go
@@ -19,7 +19,6 @@ func TestTypeConversion(t *testing.T) {
 			So(convertToGoType("Object"), ShouldEqual, "Object")
 			So(convertToGoType("string[]"), ShouldEqual, "[]string")
 			So(convertToGoType("string[][]"), ShouldEqual, "[][]string")
-			So(convertToGoType("number{}"), ShouldEqual, "map[string]float64")
 			So(convertToGoType("string | Person"), ShouldEqual, "interface{}")
 			So(convertToGoType("(string | Person)[]"), ShouldEqual, "[]interface{}")
 		})
