+++ raml/samples/bad_raml.raml
@@ -33,7 +33,7 @@ title: Notes Example API
     /yet_another:
       get:
         description: Yet another
-uriParameters:
+baseUriParameters:
   communityDomain:
     displayName: Community Domain
     type: string
